requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
supabase==2.4.6
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
reportlab==4.1.0
google-generativeai==0.5.4
httpx==0.27.0
apscheduler==3.10.4
python-multipart==0.0.9
python-dotenv==1.0.1
